Consolidate plugin adapters
