--- backend/pkg/plugins/readers/gmail/guide.json
@@ -1,44 +1,0 @@
-{
-  "sections": [
-    {
-      "title": "Enable the Gmail API",
-      "steps": [
-        {"text": "Go to console.cloud.google.com and open or create a project."},
-        {"text": "Search for 'Gmail API' and click Enable."}
-      ],
-      "link": { "label": "Google Cloud Console", "url": "https://console.cloud.google.com" }
-    },
-    {
-      "title": "Create OAuth 2.0 credentials",
-      "steps": [
-        {"text": "Navigate to APIs & Services → Credentials."},
-        {"text": "Click Create Credentials → OAuth client ID."},
-        {"text": "Set Application type to Web application."},
-        {"text": "Under Authorized redirect URIs add: http://localhost:8080/api/auth/callback (replace with your server's address if not running locally)."},
-        {"text": "Click Create, then Download JSON — this is your client_secret.json."}
-      ]
-    },
-    {
-      "title": "Configure the OAuth consent screen",
-      "steps": [
-        {"text": "Set User Type to External."},
-        {"text": "Add your email address as a test user."},
-        {"text": "Add scope: https://www.googleapis.com/auth/gmail.readonly"}
-      ]
-    }
-  ],
-  "notes": [
-    {
-      "type": "warning",
-      "text": "The redirect URI must exactly match what you enter in Google Cloud Console. If you access Expensor on a non-default port or hostname, set BASE_URL in your docker-compose.yml and use that URL here."
-    },
-    {
-      "type": "tip",
-      "text": "Running on a homeserver without a public domain? Choose Desktop app instead of Web application — then use the manual URL paste flow on the authorize screen."
-    },
-    {
-      "type": "info",
-      "text": "Only gmail.readonly is required. Expensor never modifies your inbox."
-    }
-  ]
-}
