--- backend/pkg/plugins/readers/thunderbird/guide.json
@@ -1,35 +1,0 @@
-{
-  "sections": [
-    {
-      "title": "Find your Thunderbird profile",
-      "steps": [
-        {"text": "In Thunderbird, open Help → Troubleshooting Information."},
-        {
-          "text": "Under Application Basics → Profile Folder → Click \"Show in Finder\" — this is the path to enter below.",
-          "sub_steps": [
-            "macOS: ~/Library/Thunderbird/Profiles/<name>",
-            "Linux: ~/.thunderbird/<name>",
-            "Windows: %APPDATA%/Thunderbird/Profiles/<name>"
-          ]
-        }
-      ]
-    },
-    {
-      "title": "Select mailboxes to scan",
-      "steps": [
-        {"text": "Pick the mailboxes (folders) Expensor should scan for transaction emails."},
-        {"text": "INBOX is the most common choice; add others like Promotions or bank-specific folders if needed."}
-      ]
-    }
-  ],
-  "notes": [
-    {
-      "type": "docker",
-      "text": "Running Expensor in Docker? Mount your profile directory as a read-only volume in docker-compose.yml: - /path/to/your/profile:/thunderbird-profile:ro. The profile selector will then show it automatically."
-    },
-    {
-      "type": "info",
-      "text": "If no profiles appear in the dropdown, the profile path may not be accessible. Enter it manually or check the Docker volume mount."
-    }
-  ]
-}
